agregando permisos en vistas de garantias

+++ app/views/warranty/partials/form_destroy.blade.php
@@ -1,8 +1,10 @@
-{{ Form::open(['route' => ['warranty.destroy', $warranty->id], 'method' => 'delete', 'role' => 'form', 'id' => 'form-delete']) }}
+@if(p(121))
+    {{ Form::open(['route' => ['warranty.destroy', $warranty->id], 'method' => 'delete', 'role' => 'form', 'id' => 'form-delete']) }}
 
-<button type="submit" class="btn-red" title="Eliminar garantía">
-    <i class="fa fa-times"></i>
-    Eliminar garantía {{ $warranty->folio }}
-</button>
+    <button type="submit" class="btn-red" title="Eliminar garantía">
+        <i class="fa fa-times"></i>
+        Eliminar garantía {{ $warranty->folio }}
+    </button>
 
-{{ Form::close() }}
+    {{ Form::close() }}
+@endif
