+++ app/views/warranty/partials/form_destroy_float.blade.php
@@ -1,6 +1,8 @@
-<div class="confirm-dialog hide" title="Eliminar garantía" id="dialogConfirm" data-width="400">
-    <div class="mesasge text-center">
-        <h3>¿Estas seguro de querer eliminar la garantía <span class="data_name"></span>?</h3>
+@if(p(121))
+    <div class="confirm-dialog hide" title="Eliminar garantía" id="dialogConfirm" data-width="400">
+        <div class="mesasge text-center">
+            <h3>¿Estas seguro de querer eliminar la garantía <span class="data_name"></span>?</h3>
+        </div>
     </div>
-</div>
-{{ Form::open(['route' => ['warranty.destroy', 'PRODUCT_ID'], 'method' => 'delete', 'role' => 'form', 'id' => 'form-delete']) }}{{ Form::close() }}
+    {{ Form::open(['route' => ['warranty.destroy', 'PRODUCT_ID'], 'method' => 'delete', 'role' => 'form', 'id' => 'form-delete']) }}{{ Form::close() }}
+@endif
